2) separate tests
